+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 #### Added
 
 #### Changed
+- #9945 Change text in the panel with Analytics
 
 #### Removed