+++ CHANGELOG.md
@@ -3,8 +3,10 @@
 ## UNRELEASED
 
 #### Added
+- #9951 Total event loss 
 
 #### Changed
+- #10008 Fix layer config for old events.
 - #9945 Change text in the panel with Analytics
 
 #### Removed